Customer-Debts-DisplayDebtsWithNextPaymentDueDate-Script
Created script DisplayDebtsWithNextPaymentDueDate to return all debt objects with the next payment due date field. The script has a method displayDebtsWithNextPaymentDueDate that takes a list of debt objects that  have the remaining balance field on them, and a custom dict structure I created to hold the payment plan per debtId.